Named server and deprecation changes
NERSCSlurmSpawner already extends base's `get_env` instead of creating its own, so not all of the changes made in the [corresponding PR for SSHSpawner](https://github.com/krinsman/sshspawner/commit/c0d479bc064e8647f990048ebe0e93088d830809) are necessary. 

That being said (I haven't tested it, so I don't know for certain), I imagine that the attempt to set `JPY_COOKIE_NAME` would be as much of an issue for this spawner when using [the named server implementation](https://github.com/jupyterhub/jupyterhub/pull/2089) as it was for SSHSpawner.

The setting of the other deleted environment variables is made redundant by the fact that the non-deprecated counterparts of the variables [are already set properly in the parent's `get_env` implementation](https://github.com/jupyterhub/jupyterhub/blob/master/jupyterhub/spawner.py#L609), so I removed those lines as well.